+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unrealeased
 Core
 ~~~~
 
+- API: Return basket customer information
 - API: Add option to reset password for authenticated users
 - API: Add opiton to reset password with token
 - API: `shuup.front.apps.auth.forms.PasswordRecoveryForm` to core